pyproject.toml
[project]
name = "ceadar-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.4",
    "docx2txt>=0.9",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.4",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "langserve>=0.3.3",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "unstructured>=0.18.18",
    "uvicorn>=0.38.0",
]
